# compute_node/frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Samurai Dashboard",
    "description": "REST + WebSocket API для управления роботом.",
    "version": "2.0"
  },
  "paths": {
...
    "/api/v1/patrol/command": {
      "post": {
        "tags": [
          "control"
        ],
        "summary": "Patrol Command",
        "operationId": "patrol_command_api_v1_patrol_command_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatrolCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAck"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/patrol/waypoints": {
      "post": {
        "tags": [
          "control"
        ],
        "summary": "Patrol Waypoints",
        "description": "Установить список waypoints для patrol-маршрута.",
        "operationId": "patrol_waypoints_api_v1_patrol_waypoints_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatrolWaypointsCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAck"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/path_recorder/command": {
      "post": {
        "tags": [
          "control"
        ],
        "summary": "Path Recorder Command",
        "operationId": "path_recorder_command_api_v1_path_recorder_command_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PathRecorderCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAck"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/precision_drive/command": {
      "post": {
        "tags": [
          "control"
        ],
        "summary": "Precision Drive Command",
        "description": "Сценарии точного драйвинга (cross/square/line/zigzag/goto).",
        "operationId": "precision_drive_command_api_v1_precision_drive_command_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrecisionDriveCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAck"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/calibration/command": {
      "post": {
        "tags": [
          "control"
        ],
        "summary": "Calibration Command",
        "description": "Admin: start/stop/reset калибровки.",
        "operationId": "calibration_command_api_v1_calibration_command_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalibrationCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAck"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/calibration/set": {
      "post": {
        "tags": [
          "control"
        ],
        "summary": "Calibration Set",
        "description": "Установить scale_fwd/bwd/motor_trim напрямую.",
        "operationId": "calibration_set_api_v1_calibration_set_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalibrationSetCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAck"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/calibration/profile/load": {
      "post": {
        "tags": [
          "control"
        ],
        "summary": "Calibration Profile Load",
        "operationId": "calibration_profile_load_api_v1_calibration_profile_load_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalibrationProfileLoadCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAck"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/calibration/profile/save": {
      "post": {
        "tags": [
          "control"
        ],
        "summary": "Calibration Profile Save",
        "operationId": "calibration_profile_save_api_v1_calibration_profile_save_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalibrationProfileSaveCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAck"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/calibration/profile/delete": {
      "post": {
        "tags": [
          "control"
        ],
        "summary": "Calibration Profile Delete",
        "operationId": "calibration_profile_delete_api_v1_calibration_profile_delete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalibrationProfileDeleteCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAck"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/mission/command": {
      "post": {
        "tags": [
          "control"
        ],
        "summary": "Mission Command",
        "operationId": "mission_command_api_v1_mission_command_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MissionCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAck"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/explorer/command": {
      "post": {
        "tags": [
          "control"
        ],
        "summary": "Explorer Command",
        "operationId": "explorer_command_api_v1_explorer_command_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExplorerCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandAck"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CalibrationCommand": {
        "properties": {
          "command": {
            "type": "string",
            "enum": [
              "start",
              "stop",
              "reset",
              "pause",
              "resume"
            ],
            "title": "Command"
          }
        },
        "type": "object",
        "required": [
          "command"
        ],
        "title": "CalibrationCommand",
        "description": "POST /api/calibration/command — admin string command (start/stop/reset)."
      },
...
      "CalibrationProfileDeleteCommand": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_\\-]+$",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CalibrationProfileDeleteCommand",
        "description": "POST /api/calibration/profile/delete."
      },
...
      "CalibrationProfileLoadCommand": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_\\-]+$",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CalibrationProfileLoadCommand",
        "description": "POST /api/calibration/profile/load."
      },
      "CalibrationProfileSaveCommand": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[a-zA-Z0-9_\\-]+$",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CalibrationProfileSaveCommand",
        "description": "POST /api/calibration/profile/save."
      },
      "CalibrationSetCommand": {
        "properties": {
          "scale_fwd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scale Fwd"
          },
          "scale_bwd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scale Bwd"
          },
          "motor_trim": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Motor Trim"
          }
        },
        "type": "object",
        "title": "CalibrationSetCommand",
        "description": "POST /api/calibration/set."
      },
...
      "CommandAck": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          }
        },
        "type": "object",
        "title": "CommandAck",
        "description": "Подтверждение команды (без возвращаемых данных)."
      },
...
      "ExplorerCommand": {
        "properties": {
          "command": {
            "type": "string",
            "enum": [
              "start",
              "stop"
            ],
            "title": "Command"
          },
          "strategy": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "frontier",
                  "spiral",
                  "zigzag"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Strategy",
            "default": "frontier"
          }
        },
        "type": "object",
        "required": [
          "command"
        ],
        "title": "ExplorerCommand",
        "description": "POST /api/explorer/command."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MissionCommand": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "args": {
            "additionalProperties": true,
            "type": "object",
            "title": "Args"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "MissionCommand",
        "description": "POST /api/mission/command — высокоуровневые сценарии."
      },
...
      "PathRecorderCommand": {
        "properties": {
          "command": {
            "type": "string",
            "enum": [
              "record",
              "stop",
              "replay",
              "pause",
              "resume"
            ],
            "title": "Command"
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "pattern": "^[a-zA-Z0-9_\\-]*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "Имя пути для load/save (только safe chars)"
          }
        },
        "type": "object",
        "required": [
          "command"
        ],
        "title": "PathRecorderCommand",
        "description": "POST /api/path_recorder/command."
      },
...
      "PatrolCommand": {
        "properties": {
          "command": {
            "type": "string",
            "enum": [
              "start",
              "stop",
              "pause",
              "resume"
            ],
            "title": "Command"
          }
        },
        "type": "object",
        "required": [
          "command"
        ],
        "title": "PatrolCommand",
        "description": "POST /api/patrol/command."
      },
...
      "PatrolWaypointsCommand": {
        "properties": {
          "waypoints": {
            "items": {
              "$ref": "#/components/schemas/PatrolWaypoint"
            },
            "type": "array",
            "title": "Waypoints"
          }
        },
        "type": "object",
        "required": [
          "waypoints"
        ],
        "title": "PatrolWaypointsCommand",
        "description": "POST /api/patrol/waypoints."
      },
...
      "PrecisionDriveCommand": {
        "properties": {
          "scenario": {
            "type": "string",
            "enum": [
              "cross",
              "square",
              "line",
              "zigzag",
              "goto"
            ],
            "title": "Scenario"
          },
          "target_x": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target X"
          },
          "target_y": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Y"
          },
          "target_theta": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Theta"
          },
          "distance_cm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Distance Cm"
          }
        },
        "type": "object",
        "required": [
          "scenario"
        ],
        "title": "PrecisionDriveCommand",
        "description": "POST /api/precision_drive/command."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}